Fold high-contention benchmark into snake_bench_test.go
Move BenchmarkSnake_HighContention and its runPooledContention/
contentionLevels helpers out of the standalone snake_contention_bench_test.go
and into the canonical snake_bench_test.go, consolidating the loadshed
package's benchmarks in one file. Pure relocation; benchmark logic is
unchanged.

AI disclosure: Claude Code assisted with development. Every line of code was either written by or carefully reviewed by me :)
